[CHORE]: type 수정
