+++ src/actions/chatbot/chat.ts
@@ -1,3 +1,4 @@
+import { chatbotAnswerType } from "@/types/chatbotType";
 import { requestWikoAI } from "../common/common";
 
 export async function postChatting(
@@ -15,14 +16,4 @@ export async function postChatting(
     console.error("챗봇 응답 에러", error);
     throw new Error(`챗봇 응답 에러 ${error}`);
   }
-}
-
-export interface chatbotAnswerType {
-  answer: string;
-  sub_questions: string[] | null | centerDataType;
-}
-export interface centerDataType {
-  center_name: string;
-  address: string;
-  telephone: string;
 }
